14.04.11

+++ src/main/java/com/example/testapp/serviceimplement/AvisService.java
@@ -8,17 +8,24 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
 @Service
 public class AvisService implements AvisInterface {
+
     @Autowired
-    AvisRepository avisRepository;
+    private AvisRepository avisRepository;
+
     @Override
     public Avis addAvis(Avis avis) {
         return avisRepository.save(avis);
     }
 
+
     @Override
     public void deleteAvis(Long id) {
+        if (!avisRepository.existsById(id)) {
+            throw new EntityNotFoundException("Avis non trouvé avec l'ID : " + id);
+        }
         avisRepository.deleteById(id);
     }
 
@@ -34,9 +41,11 @@ public class AvisService implements AvisInterface {
 
     @Override
     public Avis updateAvis(Long id, Avis avis) {
-        Avis avis1 = avisRepository.findById(id).orElseThrow(()-> new EntityNotFoundException("Avis non trouver"));
-        avis1.setCommentaireAvis(avis.getCommentaireAvis());
-        return avisRepository.save(avis1);
+        Avis avisExistant = avisRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Avis non trouvé avec l'ID : " + id));
+
+        avisExistant.setCommentaireAvis(avis.getCommentaireAvis());
+        return avisRepository.save(avisExistant);
     }
 
     @Override
